+++ BetterChests/CHANGELOG.md
@@ -3,9 +3,14 @@
 ## 1.2.0 (Unreleased)
 
 * The ItemSelectionMenu is now affected by Better Chest features.
-  * ResizeChestMenu will expand the ItemSelectionMenu.
-  * ChestMenuTabs will filter the ItemSelectionMenu.
-  * SearchItems will add the search bar to the ItemSelectionMenu.
+    * ResizeChestMenu will expand the ItemSelectionMenu.
+    * ChestMenuTabs will filter the ItemSelectionMenu.
+    * SearchItems will add the search bar to the ItemSelectionMenu.
+* Added CarryChestLimit option to limit the number of chests that can be carried at once.
+* Added red text alerts to certain features.
+  * When CarryChestLimit is reached and attempting to carry another chest.
+  * When attempting to Craft from Chests and no eligible chests were found.
+  * When attempting to Stash to Chests and no eligible chests were found.
 
 ## 1.1.0 (February 5, 2022)